pyproject.toml
[tool.poetry]
name = "connect-telegram-bot"
version = "0.1.0"
description = "Connect Telegram Bot"
authors = ["CloudBlue LLC"]
license = "Apache-2.0"
packages = [
    { include = "connect" },
]
readme = "./README.md"
homepage = "https://connect.cloudblue.com"
repository = "https://github.com/cloudblue/connect-telegram-bot"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries",
]

keywords = [
    "telegram",
    "connect",
    "utility",
    "cloudblue",
    "bot",
]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.3"
django = "2.2.19"

[tool.poetry.dev-dependencies]
ipython = "^7.21.0"
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
coverage = {extras = ["toml"], version = "^5.3"}
flake8 = "~3.8"
flake8-bugbear = "~20"
flake8-cognitive-complexity = "^0.1"
flake8-commas = "~2.0"
flake8-future-import = "~0.4"
flake8-import-order = "~0.18"
flake8-broken-line = "~0.3"
fs = "^2.4.12"
django-fake-database-backends = "0.1.1"
django-fake-model = "0.1.4"
django-mock-queries = "2.1.6"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"
log_cli = true
addopts = "--cov=connect.telegram_bot --cov-report=term-missing:skip-covered --cov-report=html --cov-report=xml"

[tool.coverage.run]
branch = true

[tool.coverage.report]
omit = [
]

exclude_lines = [
    "pragma: no cover",
    "def __str__",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]